quick nit
