+++ web/src/app/ee/admin/billing/InfoItem.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+interface InfoItemProps {
+  title: string;
+  value: string;
+}
+
+export function InfoItem({ title, value }: InfoItemProps) {
+  return (
+    <div className="bg-muted p-4 rounded-lg">
+      <p className="text-sm font-medium text-muted-foreground mb-1">{title}</p>
+      <p className="text-lg font-semibold text-foreground dark:text-white">
+        {value}
+      </p>
+    </div>
+  );
+}
